global refactor - renamed and replaced components in new directories - TableComponent folder
