+++ src/pagesMinin/Cards.page.tsx
@@ -7,8 +7,8 @@ import { PaginationWithSelect } from '@/components/ui/Pagination/PaginationWithS
 import Typography from '@/components/ui/Typography/Typography'
 import { Button } from '@/components/ui/button'
 import { ModalAddEditCard } from '@/pagesMinin/ModalsForTable/ModalEditCard/ModalAddEditCard'
-import { SingleRowCard } from '@/pagesMinin/Tables/CardsTable/SingleRowCard'
-import { TableComponent } from '@/pagesMinin/Tables/TableComponent'
+import { SingleRowCard } from '@/pagesMinin/TableComponent/SingleRowCard/SingleRowCard'
+import { TableComponent } from '@/pagesMinin/TableComponent/TableComponent'
 import { Page } from '@/pagesMinin/componentsMinin/Page/Page'
 import { useQueryParams } from '@/pagesMinin/utls/useQueryParams'
 import {
@@ -118,7 +118,7 @@ export const CardsPage = () => {
         </div>
       ) : (
         <>
-          {/*<CardsTable data={data} tableHeader={headersNameCards} />*/}
+          {/*<SingleRowCard data={data} tableHeader={headersNameCards} />*/}
           <TableComponent data={data} tableHeader={headersNameCards}>
             {item => <SingleRowCard item={item} />}
           </TableComponent>
