+++ src/pagesMinin/TableComponent/SingleRowCard/CardsTable.module.scss
