+++ src/pagesMinin/TableComponent/SingleRowDeck/decksTable.module.scss
