+++ src/pagesMinin/TableComponent/tableComponent.module.scss
