+++ src/pagesMinin/TableComponent/tableSingleRow.module.scss
